create separate component for just ComingSoonTitle and modify animation slightly
