+++ src/app/views/coming-soon/ComingSoon.scss
@@ -46,139 +46,4 @@ body {
 		margin-top: 25vh;
 		background: url($coral) no-repeat top/cover;
 	}
-
-	.coming-soon-text {
-		text-align: center;
-		z-index: 10;
-
-		#coming-soon-title-container {
-			display: flex;
-			justify-content: center;
-			gap: 1.3rem;
-			font-family: "Londrina Solid", sans-serif;
-			font-weight: 400;
-			font-size: 9rem;
-			color: white;
-			text-shadow: 4px 4px 6px rgba(0, 0, 0, 0.5), 9px 16px 10px #68b3d7;
-
-			@media screen and (max-width: vh-styles.$break-medium) {
-				font-size: 6rem;
-				gap: 1rem;
-			}
-
-			@media screen and (max-width: vh-styles.$break-small) {
-				font-size: 3.5rem;
-				gap: 0.5rem;
-			}
-		}
-
-		.coming-soon-letter {
-			display: inline-block;
-			position: relative;
-		}
-
-		// floating animation: Alternate letters go up/down
-		.float-up {
-			animation: floatUp 3s ease-in-out infinite;
-		}
-
-		.float-down {
-			animation: floatDown 3s ease-in-out infinite;
-		}
-
-		@keyframes floatUp {
-			0% {
-				transform: translateY(-15px) rotate(5deg);
-			}
-			50% {
-				transform: translateY(15px) rotate(5deg);
-			}
-			100% {
-				transform: translateY(-15px) rotate(5deg);
-			}
-		}
-
-		@keyframes floatDown {
-			0% {
-				transform: translateY(15px) rotate(-5deg);
-			}
-			50% {
-				transform: translateY(-15px) rotate(-5deg);
-			}
-			100% {
-				transform: translateY(15px) rotate(-5deg);
-			}
-		}
-
-		@media screen and (max-width: vh-styles.$break-medium) {
-			@keyframes floatUp {
-				0% {
-					transform: translateY(-8px) rotate(5deg);
-				}
-				50% {
-					transform: translateY(8px) rotate(5deg);
-				}
-				100% {
-					transform: translateY(-8px) rotate(5deg);
-				}
-			}
-
-			@keyframes floatDown {
-				0% {
-					transform: translateY(8px) rotate(-5deg);
-				}
-				50% {
-					transform: translateY(-8px) rotate(-5deg);
-				}
-				100% {
-					transform: translateY(8px) rotate(-5deg);
-				}
-			}
-		}
-
-		@media screen and (max-width: vh-styles.$break-small) {
-			@keyframes floatUp {
-				0% {
-					transform: translateY(-5px) rotate(5deg);
-				}
-				50% {
-					transform: translateY(5px) rotate(5deg);
-				}
-				100% {
-					transform: translateY(-5px) rotate(5deg);
-				}
-			}
-
-			@keyframes floatDown {
-				0% {
-					transform: translateY(5px) rotate(-5deg);
-				}
-				50% {
-					transform: translateY(-5px) rotate(-5deg);
-				}
-				100% {
-					transform: translateY(5px) rotate(-5deg);
-				}
-			}
-		}
-
-		#coming-soon-subtitle {
-			font-family: "Instrument Sans", serif;
-			font-weight: 600;
-			font-size: 2.25rem;
-			color: white;
-			margin-top: -2rem;
-			text-shadow: 0px 6px 6px #69b3d7;
-
-			@media screen and (max-width: vh-styles.$break-medium) {
-				font-size: 2rem;
-				margin-top: -1.5rem;
-			}
-
-			@media screen and (max-width: vh-styles.$break-small) {
-				font-size: 1.5rem;
-				margin-top: -0.5rem;
-			}
-		}
-	}
 }
